+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - **[Fix]** Fix panic on invalid CSM text settings.
 - **[Fix]** Fix panic on incomplete clip action string.
 - **[Fix]** Fix panic on failed image dimension detection.
+- **[Fix]** Fix panic on invalid button cond action string.
 
 ## Typescript